Fixed Log Level None actually resulting in everything being logged.
Added an instance ID that is printed in each message, and at startup, so we can tell which messages happened when.
The time (mm/dd/yyyy hh:mm:ss am/pm) is now also printed with each message to give context on when something occurred.

Added an option to truncate the log on startup. Enabling truncation will also disable writing the instance ID, as the file will be cleared when the game begins, and there will be no other instance to worry about.

Signed-off-by: Drew Ritter <dritter@razelon.com>